+++ api/src/certification/enrolment/application/subscription-controller.js
@@ -1,5 +1,5 @@
 import { usecases } from '../../../../lib/domain/usecases/index.js';
-import * as certificationCandidateSubscriptionSerializer from '../../../../lib/infrastructure/serializers/jsonapi/certification-candidate-subscription-serializer.js';
+import * as certificationCandidateSubscriptionSerializer from '../infrastructure/serializers/certification-candidate-subscription-serializer.js';
 
 const getSubscription = async function (request, _h, dependencies = { certificationCandidateSubscriptionSerializer }) {
   const certificationCandidateId = request.params.id;
